reorganize tests for coverage omission
